CliOptions moved to the separate file
