+++ compiler/plugin-api/src/org/jetbrains/kotlin/compiler/plugin/CommandLineProcessor.kt
@@ -17,17 +17,6 @@
 package org.jetbrains.kotlin.compiler.plugin
 
 import org.jetbrains.jet.config.CompilerConfiguration
-import java.util.regex.Pattern
-
-public class CliOption(
-        public val name: String,
-        public val valueDescription: String,
-        public val description: String,
-        public val required: Boolean = true,
-        public val allowMultipleOccurrences: Boolean = false
-)
-
-public class CliOptionProcessingException(message: String, cause: Throwable? = null): RuntimeException(message, cause)
 
 public trait CommandLineProcessor {
     public val pluginId: String
@@ -35,26 +24,4 @@ public trait CommandLineProcessor {
 
     [throws(javaClass<CliOptionProcessingException>())]
     public fun processOption(option: CliOption, value: String, configuration: CompilerConfiguration)
-}
-
-public data class PluginOptionValue(
-        val pluginId: String,
-        val optionName: String,
-        val value: String
-) {
-    override fun toString() = "$pluginId:$optionName=$value"
-}
-
-public fun parsePluginOption(argumentValue: String): PluginOptionValue? {
-    val pattern = Pattern.compile("""^plugin:([^:]*):([^=]*)=(.*)$""")
-    val matcher = pattern.matcher(argumentValue)
-    if (matcher.matches()) {
-        return PluginOptionValue(matcher.group(1), matcher.group(2), matcher.group(3))
-    }
-
-    return null
-}
-
-public fun getPluginOptionString(pluginId: String, key: String, value: String): String {
-    return "plugin:$pluginId:$key=$value"
 }
